+++ libs/lwrb/src/include/lwrb/lwrb.h
@@ -29,13 +29,13 @@
  * This file is part of LwRB - Lightweight ring buffer library.
  *
  * Author:          Tilen MAJERLE <tilen@majerle.eu>
- * Version:         v1.3.1
+ * Version:         v3.0.0-rc1
  */
 #ifndef LWRB_HDR_H
 #define LWRB_HDR_H
 
-#include <string.h>
 #include <stdint.h>
+#include <string.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -47,25 +47,20 @@ extern "C" {
  * \{
  */
 
-/**
- * \brief           Enable buffer structure pointer parameter as volatile
- * To use this feature, uncomment keyword below
- */
-#define LWRB_VOLATILE                           /* volatile */
-
-/**
- * \brief           Adds 2 magic words to make sure if memory is corrupted
- *                  application can detect wrong data pointer and maximum size
- */
-#define LWRB_USE_MAGIC                      1
+#ifdef LWRB_DISABLE_ATOMIC
+typedef unsigned long lwrb_ulong_t;
+#else
+#include <stdatomic.h>
+typedef atomic_ulong lwrb_ulong_t;
+#endif
 
 /**
  * \brief           Event type for buffer operations
  */
 typedef enum {
-    LWRB_EVT_READ,                              /*!< Read event */
-    LWRB_EVT_WRITE,                             /*!< Write event */
-    LWRB_EVT_RESET,                             /*!< Reset event */
+    LWRB_EVT_READ,  /*!< Read event */
+    LWRB_EVT_WRITE, /*!< Write event */
+    LWRB_EVT_RESET, /*!< Reset event */
 } lwrb_evt_type_t;
 
 /**
@@ -79,50 +74,46 @@ struct lwrb;
  * \param[in]       evt: Event type
  * \param[in]       bp: Number of bytes written or read (when used), depends on event type
  */
-typedef void (*lwrb_evt_fn)(LWRB_VOLATILE struct lwrb* buff, lwrb_evt_type_t evt, size_t bp);
+typedef void (*lwrb_evt_fn)(struct lwrb* buff, lwrb_evt_type_t evt, size_t bp);
 
 /**
  * \brief           Buffer structure
  */
 typedef struct lwrb {
-#if LWRB_USE_MAGIC
-    uint32_t magic1;                            /*!< Magic 1 word */
-#endif /* LWRB_USE_MAGIC */
-    uint8_t* buff;                              /*!< Pointer to buffer data.
-                                                    Buffer is considered initialized when `buff != NULL` and `size > 0` */
-    size_t size;                                /*!< Size of buffer data. Size of actual buffer is `1` byte less than value holds */
-    size_t r;                                   /*!< Next read pointer. Buffer is considered empty when `r == w` and full when `w == r - 1` */
-    size_t w;                                   /*!< Next write pointer. Buffer is considered empty when `r == w` and full when `w == r - 1` */
-    lwrb_evt_fn evt_fn;                         /*!< Pointer to event callback function */
-#if LWRB_USE_MAGIC
-    uint32_t magic2;                            /*!< Magic 2 word */
-#endif /* LWRB_USE_MAGIC */
+    uint8_t* buff;  /*!< Pointer to buffer data. Buffer is considered initialized when `buff != NULL` and `size > 0` */
+    size_t size;    /*!< Size of buffer data. Size of actual buffer is `1` byte less than value holds */
+    lwrb_ulong_t r; /*!< Next read pointer. Buffer is considered empty when `r == w` and full when `w == r - 1` */
+    lwrb_ulong_t w; /*!< Next write pointer. Buffer is considered empty when `r == w` and full when `w == r - 1` */
+    lwrb_evt_fn evt_fn; /*!< Pointer to event callback function */
 } lwrb_t;
 
-uint8_t     lwrb_init(LWRB_VOLATILE lwrb_t* buff, void* buffdata, size_t size);
-uint8_t     lwrb_is_ready(LWRB_VOLATILE lwrb_t* buff);
-void        lwrb_free(LWRB_VOLATILE lwrb_t* buff);
-void        lwrb_reset(LWRB_VOLATILE lwrb_t* buff);
-void        lwrb_set_evt_fn(LWRB_VOLATILE lwrb_t* buff, lwrb_evt_fn fn);
+uint8_t lwrb_init(lwrb_t* buff, void* buffdata, size_t size);
+uint8_t lwrb_is_ready(lwrb_t* buff);
+void lwrb_free(lwrb_t* buff);
+void lwrb_reset(lwrb_t* buff);
+void lwrb_set_evt_fn(lwrb_t* buff, lwrb_evt_fn fn);
 
 /* Read/Write functions */
-size_t      lwrb_write(LWRB_VOLATILE lwrb_t* buff, const void* data, size_t btw);
-size_t      lwrb_read(LWRB_VOLATILE lwrb_t* buff, void* data, size_t btr);
-size_t      lwrb_peek(LWRB_VOLATILE lwrb_t* buff, size_t skip_count, void* data, size_t btp);
+size_t lwrb_write(lwrb_t* buff, const void* data, size_t btw);
+size_t lwrb_read(lwrb_t* buff, void* data, size_t btr);
+size_t lwrb_peek(const lwrb_t* buff, size_t skip_count, void* data, size_t btp);
 
 /* Buffer size information */
-size_t      lwrb_get_free(LWRB_VOLATILE lwrb_t* buff);
-size_t      lwrb_get_full(LWRB_VOLATILE lwrb_t* buff);
+size_t lwrb_get_free(const lwrb_t* buff);
+size_t lwrb_get_full(const lwrb_t* buff);
 
 /* Read data block management */
-void*       lwrb_get_linear_block_read_address(LWRB_VOLATILE lwrb_t* buff);
-size_t      lwrb_get_linear_block_read_length(LWRB_VOLATILE lwrb_t* buff);
-size_t      lwrb_skip(LWRB_VOLATILE lwrb_t* buff, size_t len);
+void* lwrb_get_linear_block_read_address(const lwrb_t* buff);
+size_t lwrb_get_linear_block_read_length(const lwrb_t* buff);
+size_t lwrb_skip(lwrb_t* buff, size_t len);
 
 /* Write data block management */
-void*       lwrb_get_linear_block_write_address(LWRB_VOLATILE lwrb_t* buff);
-size_t      lwrb_get_linear_block_write_length(LWRB_VOLATILE lwrb_t* buff);
-size_t      lwrb_advance(LWRB_VOLATILE lwrb_t* buff, size_t len);
+void* lwrb_get_linear_block_write_address(const lwrb_t* buff);
+size_t lwrb_get_linear_block_write_length(const lwrb_t* buff);
+size_t lwrb_advance(lwrb_t* buff, size_t len);
+
+/* Search in buffer */
+uint8_t lwrb_find(const lwrb_t* buff, const void* bts, size_t len, size_t start_offset, size_t* found_idx);
 
 /**
  * \}