+++ lwpkt/src/lwpkt/lwpkt.c
@@ -130,7 +130,7 @@ prv_crc_init(lwpkt_crc_t* c) {
  * \return          \ref lwpktOK on success, member of \ref lwpktr_t otherwise
  */
 lwpktr_t
-lwpkt_init(lwpkt_t* pkt, LWRB_VOLATILE lwrb_t* tx_rb, LWRB_VOLATILE lwrb_t* rx_rb) {
+lwpkt_init(lwpkt_t* pkt, lwrb_t* tx_rb, lwrb_t* rx_rb) {
     if (pkt == NULL) {
         return lwpktERR;
     }
@@ -177,6 +177,8 @@ lwpkt_read(lwpkt_t* pkt) {
     if (!LWPKT_IS_VALID(pkt)) {
         return lwpktERR;
     }
+
+    SEND_EVT(pkt, LWPKT_EVT_PRE_READ);
 
     /* Process bytes from RX ringbuffer */
     /* Read byte by byte and go through state machine */
@@ -274,11 +276,11 @@ lwpkt_read(lwpkt_t* pkt) {
                 LWPKT_SET_STATE(pkt, LWPKT_STATE_STOP);
                 break;
             }
-#endif /* LWPKT_CFG_USE_CRC */
+#endif                                                   /* LWPKT_CFG_USE_CRC */
             case LWPKT_STATE_STOP: {
                 LWPKT_SET_STATE(pkt, LWPKT_STATE_START); /* Reset packet state */
                 if (b == LWPKT_STOP_BYTE) {
-                    res = lwpktVALID; /* Packet fully valid, take data from it */
+                    res = lwpktVALID;                    /* Packet fully valid, take data from it */
                     goto retpre;
                 } else {
                     res = lwpktERRSTOP; /* Packet is missing STOP byte! */
@@ -292,15 +294,16 @@ lwpkt_read(lwpkt_t* pkt) {
             }
         }
     }
+    if (pkt->m.state == LWPKT_STATE_START) {
+        res = lwpktWAITDATA;
+    } else {
+        res = lwpktINPROG;
+    }
+retpre:
+    SEND_EVT(pkt, LWPKT_EVT_POST_READ);
     if (e) {
         SEND_EVT(pkt, LWPKT_EVT_READ); /* Send read event */
     }
-    if (pkt->m.state == LWPKT_STATE_START) {
-        return lwpktWAITDATA;
-    }
-    return lwpktINPROG;
-retpre:
-    SEND_EVT(pkt, LWPKT_EVT_READ); /* Send read event */
     return res;
 }
 
@@ -354,6 +357,7 @@ lwpkt_write(lwpkt_t* pkt,
 #endif /* LWPKT_CFG_USE_CMD || __DOXYGEN__ */
             const void* data, size_t len) {
 #if LWPKT_CFG_USE_CRC
+    lwpktr_t res = lwpktOK;
     lwpkt_crc_t crc;
 #endif /* LWPKT_CFG_USE_CRC */
 #if LWPKT_CFG_ADDR_EXTENDED
@@ -362,14 +366,17 @@ lwpkt_write(lwpkt_t* pkt,
     size_t org_len = len;
     uint8_t b;
 
+    SEND_EVT(pkt, LWPKT_EVT_PRE_WRITE);
+
     if (!LWPKT_IS_VALID(pkt)) {
-        return lwpktERR;
+        res = lwpktERR;
+        goto fast_return;
     } else {
         /* Check for required memory for packet */
         size_t min_mem = 2, tmp_len = 0;
 #if LWPKT_CFG_USE_ADDR && LWPKT_CFG_ADDR_EXTENDED
         lwpkt_addr_t tmp_addr;
-#endif /* LWPKT_CFG_USE_ADDR && LWPKT_CFG_ADDR_EXTENDED */
+#endif  /* LWPKT_CFG_USE_ADDR && LWPKT_CFG_ADDR_EXTENDED */
 
         /* Addresses */
 #if LWPKT_CFG_USE_ADDR
@@ -411,7 +418,8 @@ lwpkt_write(lwpkt_t* pkt,
 
         /* Verify enough memory */
         if (lwrb_get_free(pkt->tx_rb) < min_mem) {
-            return lwpktERRMEM;
+            res = lwpktERRMEM;
+            goto fast_return;
         }
     }
 
@@ -472,10 +480,13 @@ lwpkt_write(lwpkt_t* pkt,
     b = LWPKT_STOP_BYTE;
     lwrb_write(pkt->tx_rb, &b, 1);
 
+fast_return:
     /* Final step to notify app */
-    SEND_EVT(pkt, LWPKT_EVT_WRITE); /* Send write event */
-
-    return lwpktOK;
+    SEND_EVT(pkt, LWPKT_EVT_POST_WRITE); /* Release write event */
+    if (res == lwpktOK) {
+        SEND_EVT(pkt, LWPKT_EVT_WRITE);  /* Send write event */
+    }
+    return res;
 }
 
 /**